backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
supabase==1.2.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.1.0
web3==6.11.3
eth-account==0.9.0
eth-utils==2.2.2
psycopg2-binary==2.9.7
cryptography==41.0.7